Fixed some bugs in rtp

+++ rtp/worker.go
@@ -104,7 +104,7 @@ func StartRTP(game sgc7game.IGame, rtp *RTP, worknums int, spinnums int64, stake
 					i--
 				} else {
 					if limitPayout > 0 {
-						cp := 0
+						cp := int64(0)
 						for _, v := range results {
 							if cp+v.CashWin > limitPayout {
 								cp += v.CashWin
